backend/python/requirements.txt
fastapi
uvicorn
boto3
pydantic
python-multipart
python-dotenv
pandas
pdfplumber
python-docx
google-genai
